refactor: refactor createStatusTree function to utils file
